requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.27
botocore==1.35.27
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
duckduckgo_search==7.2.0
edge-tts==6.1.12
exceptiongroup==1.2.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
googlesearch-python==1.2.5
gotrue==2.8.1
groq==0.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.0
langchain-core==0.3.5
langchain-groq==0.2.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.126
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
multidict==6.1.0
mutagen==1.47.0
narwhals==1.8.3
numpy==1.26.4
openai==1.47.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
postgrest==0.16.11
posthog==3.7.0
primp==0.9.3
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
realtime==2.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
sentry-sdk==2.14.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
storage3==0.7.7
streamlit==1.39.0
streamlit-cookies-controller==0.0.4
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
websockets==12.0
yarl==1.12.1